src/requirements.txt
# Core Python libraries
pandas==2.3.2         # Data manipulation
numpy==2.3.3          # Numerical computing
python-dotenv==1.1.1  # Load environment variables from .env

# HTTP / API
requests==2.32.5      # Making API requests

# PostgreSQL
psycopg2-binary==2.9.10  # PostgreSQL driver

# Jupyter / Analytics
jupyter==7.4.5        # Jupyter notebook
ipykernel==6.30.1     # IPython kernel for notebooks
notebook==7.4.5       # Notebook server
matplotlib-inline==0.1.7  # Inline plotting (optional)
